chore: refactor variable and function names to make them consistent with the latest conventions
Most important change is to drop the use of gate to mean the feeder link or the subroot or the detour node proximal to the root. Now 'feeder' is the segmented line from root to 'subroot' (always a terminal), while 'proximal' is the vertex (either clone or terminal) that is part of feeder and closest to root.