feat: add capability of loading obstacles from .osm.pbf and clever handling of root on concavity
This introduces important changes in the loading of locations. All border's and obstacles'
vertices are now checked for overlapping with node coordinates. Any overlapping vertices
result in references to the node coordinate instead of duplicating the coordinates.
This allows for not creating stunt vertices for substations so that, when a substation
coincides with a concavity vertex, there is no passage for routes between the substation
and the concavity. With this change, the avoidance of crossing with export cables can
be enforced by creating concavities that touch one or more substations.