feat: add to PathFinder option bad_streak_limit and stop traverser when it revisits a node
The revisit part can cause some reduction in detour quality in some edge cases. It may be more robust to monitor triangle revisits instead of the (node, sector) revisits that this commit implements. The reason for this change is to be more robust in poorly connected solutions as the ones created by the initial stages of data_driven_heuristic().