feat: first attempt at implementing constraints lone longest triangle
The MILP solver seems to waste time on configurations that use two sides of the same Delaunay triangle in a subtree, where one of the two sides is the longest side of the triangle. This configuration can always be improved by switching from the longest side to the unused side.

Taking the binary variables representing each triangle's sides: `x_short_1`, `x_short_2`, `x_long` in A.edges, the constraints are 2 per triangle: `x_short_1 + x_long <= 1` and `x_short_2 + x_long <= 1`.