feat: add FiberSCIP (fscip) solver (wip, need to parse multiple-solution output)
fscip is writing all solutions to the solution file best.sol, which is not expected by pyscipopt to have multiple solutions. This file needs to be parsed and split in multiple files (there is a blank line between solutions, except for the last one, where the dividing line has "[ Final Solution ]".
On top of that, python is mysteriously halting when subprocess.run(fscip) returns.